Revert "Navigaate records in form"
