+++ app/graphql/types/query_type.rb
@@ -14,25 +14,6 @@ module Types
       argument :batch, Int, required: true
     end
 
-    field :last_calibration_by_batch, Types::CalibrationType, null: false do
-      argument :batch, Int, required: false
-    end
-
-    field :previous_calibration, Types::CalibrationType, null: true do
-      argument :batch, Int, required: false
-      argument :id, ID, required: false
-    end
-
-    field :next_calibration, Types::CalibrationType, null: true do
-      argument :batch, Int, required: false
-      argument :id, ID, required: false
-    end
-
-    field :dosimeter_by_batch, Types::DosimeterType, null: false do
-      argument :batch, Int, required: false
-      argument :id, ID, required: false
-    end
-
     field :customer_by_batch, Types::CustomerType, null: true do
       argument :batch, Int, required: false
     end
@@ -62,33 +43,7 @@ module Types
     end
 
     def calibrations_by_batch(batch:)
-      Calibration.where(batch: batch)
-    end
-
-    def last_calibration_by_batch(batch:)
-      Calibration.where(batch: batch).last
-    end
-
-    def previous_calibration(batch:, id:)
-      if id
-        Calibration.prev(batch, id)
-      else
-        Calibration.where(batch: batch).last
-        # last_id = Calibration.last.id
-        # Calibration.where("batch =  ? AND id = ?", batch, last_id - 1).first
-      end
-    end
-
-    def next_calibration(batch:, id:)
-      Calibration.next(batch, id)
-    end
-
-    def dosimeter_by_batch(batch:, id:)
-      if id
-        Calibration.find(id).dosimeter
-      else
-        Calibration.where(batch: batch).last.dosimeter
-      end
+      calibrations = Calibration.where(batch: batch)
     end
 
     def customer_by_batch(batch:)
