+++ app/models/calibration.rb
@@ -2,13 +2,4 @@ class Calibration < ApplicationRecord
   belongs_to :dosimeter
   has_many :customer, through: :dosimeters
   belongs_to :user
-
-  def self.prev(batch, cal_id)
-    Calibration.where("batch =  ? and id < ?", batch, cal_id).last
-  end
-
-  def self.next(batch, cal_id)
-    Calibration.where("batch =  ? and id > ?", batch, cal_id).first
-  end
-  
 end
